Remove outdated docker scripts, add new entity with repository
Deleted outdated docker file and configuration as they are no longer being used in the new software architecture. Decided to go with a new DragonTreasure entity added to support the recent feature of user collection of treasures. This change also includes auto-generated DB migration for the entity. New entity is supported by some extra Symfony features added to the project.